Done some cleaning
