+++ TPV_Source/configs/app_config.ini
@@ -15,4 +15,4 @@ text1 = textfile1
 text2 = textfile2
 
 [Diversje]
-1 = 884x502
+1 = 861x545
