+++ TPV_Source/configs/config1.ini
@@ -50,7 +50,7 @@
 5 = "TPV Skjema for Maskin 20, Rev.1"
 
 [Filbehandling]
-1 = C:/Users/heiv085/Documents/Github/TPV_Skjema/TPV_Source/test.xlsx
+1 = /Users/iverullrikhepso/Documents/test.xlsx
 2 = 837x475
 3 = ""
 4 = C:/Users/Iver/Documents/test.docx
