+++ TPV_Source/configs/config2.ini
@@ -37,7 +37,7 @@
 1 = C:/Users/heiv085/Documents/Github/TPV_Skjema/TPV_Source/test.xlsx
 2 = 746x387
 3 = ""
-4 = ""
+4 = /Users/iverullrikhepso/Documents/Vedlikehold.docx
 [Vedlikeholdsjekk]
 1 = ""
 2 = ""
